+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added XDG_CURRENT_DESKTOP environment
 - added XAUTHORITY environment
 - changed Nix runtime definition to contain a single installable
+- added support for local flake usage
 
 ## 0.6.0 (2024-10-16)